feat: agregando exceptions más limpias y con cuerpo JSON, y probando VehiculoDTO

+++ vehiculos-api/src/main/java/com/proy/vehiculos_api/exception/VehicleNotFoundException.java
@@ -2,7 +2,7 @@ package com.proy.vehiculos_api.exception;
 
 public class VehicleNotFoundException extends RuntimeException {
 	
-	public VehicleNotFoundException(int id) {
+	public VehicleNotFoundException(Long id) {
 		super("Vehiculo con id " + id + " no encontrado");
 	}
